Reuse identical KeyValues across traces in a SearchPage to reduce final binary size
